requirements.txt
python-dateutil
flask
gunicorn
pmdarima
statsmodels
pandas
numpy